requirements.txt
pytelegrambotapi
aiogram>=3.0.0
pandas
streamlit
supabase
python-dotenv
openai>=1.0.0
# gigachat>=0.1.20  # Закомментировано - перешли на OpenAI
qrcode[pil]
pyzbar
fastapi
python-multipart
uvicorn[standard]>=0.30.0
sentry-sdk[fastapi]>=2.0.0
slowapi>=0.1.9
requests>=2.31.0